Member 객체의 id 타입 변경 및 store 제네릭 타입 변경
id타입은 Long에서 int로 변경
store는 map에서 list로 변경

+++ src/main/java/com/kakao/cafe/repository/MemberRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 public interface MemberRepository {
     Member save(Member member);
 
-    Optional<Member> findById(Long id);
+    Optional<Member> findById(int id);
 
     Optional<Member> findByName(String name);
